requirements.txt
mpi4py>=3.1
openai-whisper>=20240930
PyYAML>=6.0
numpy>=1.24
openai>=1.40
scikit-learn>=1.3
jinja2>=3.1
pandas>=1.5

